Move TestFixtures to top of package
