+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/utils/condition/DoubleClickTest.java
@@ -1,6 +1,6 @@
 package com.github.khanshoaib3.minecraft_access.utils.condition;
 
-import com.github.khanshoaib3.minecraft_access.utils.TestFixtures;
+import com.github.khanshoaib3.minecraft_access.TestFixtures;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
